% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isoreg.R
\name{isoreg_with_xgboost}
\alias{isoreg_with_xgboost}
\title{Perform Univariate Isotonic Regression using XGBoost}
\usage{
isoreg_with_xgboost(x, y, max_depth = 12, min_child_weight = 20)
}
\arguments{
\item{x}{A numeric vector or one-column matrix containing the predictor values.}

\item{y}{A numeric vector containing the response variable.}

\item{max_depth}{The maximum depth of the trees. Default is 15.}

\item{min_child_weight}{Minimum sum of instance weight (hessian) needed in a child.
Default is 10.
For equal weights, this is the minimal number of observations in each leaf node.}
}
\value{
A function that takes a vector or one-column matrix of predictor values
        and returns a function that maps input to the predictions fit using isotonic regression.
}
\description{
This function uses the XGBoost algorithm to perform isotonic regression
on univariate data.
}
\examples{
# Generate some example data
x <- seq(1, 100, by = 1)
y <- sort(runif(100))

# Fit the isotonic model
isotonic_function <- isoreg_with_xgboost(x, y)

# Use the fitted model to make predictions
predicted_values <- isotonic_function(x)

}
\seealso{
\code{\link[xgboost::xgb.train]{xgb.train}}
}
